Release v0.2.0: Convert binary sensors to enum sensors
