+++ custom_components/huckleberry/manifest.json
@@ -7,5 +7,5 @@
   "issue_tracker": "https://github.com/Woyken/huckleberry-homeassistant/issues",
   "requirements": ["huckleberry-api>=0.1.12"],
   "iot_class": "cloud_push",
-  "version": "0.1.14"
+  "version": "0.2.0"
 }
