+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "huckleberry-homeassistant"
-version = "0.1.14"
+version = "0.2.0"
 description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.11"
